Dev to Master

+++ tcp-server/src/main/java/ru/protei/server/service/WordService.java
@@ -1,12 +1,15 @@
 package ru.protei.server.service;
 
+import org.apache.log4j.Logger;
 import ru.protei.server.dao.WordDAO;
 import ru.protei.server.model.Word;
 
 import java.util.List;
 
 public class WordService {
-    private static WordService instance = null;
+    private static final Logger log = Logger.getLogger(WordService.class);
+
+    private static WordService instance;
     private WordDAO wordDAO;
 
     private WordService() {
@@ -14,28 +17,67 @@ public class WordService {
     }
 
     public static WordService getInstance() {
+
         if (instance == null) {
             instance = new WordService();
         }
+
         return instance;
     }
 
-    public Word find(String title) {
-        return wordDAO.find(title);
+    public Word find(Word w) {
+        return wordDAO.find(w);
     }
+
     public List<Word> findAll(String mask) {
         return wordDAO.findAll(mask);
     }
-    public Word create(Word w) {
-        return wordDAO.create(w);
+
+    public boolean create(Word w) {
+
+        boolean result = wordDAO.create(w);
+
+        if (result) {
+            log.info("Insert WORD value");
+        } else {
+            log.info("Can't insert WORD value");
+        }
+
+        return result;
     }
-    public Word update(Word w) {
-        return wordDAO.update(w);
+
+    public boolean update(Word w) {
+
+        boolean result = wordDAO.update(w);
+
+        if (result) {
+            log.info("Update WORD value");
+        } else {
+            log.info("Can't update WORD value");
+        }
+
+        return result;
     }
+
     public boolean delete(Word w) {
-        return wordDAO.delete(w);
+
+        if (!wordDAO.exists(w)) {
+            log.info("Can't delete WORD value. It doesn't exist");
+            return false;
+        }
+
+        boolean result = wordDAO.delete(w);
+
+        if (result) {
+            log.info("Delete WORD value");
+        } else {
+            log.info("Can't delete WORD value");
+        }
+
+        return result;
     }
-    public boolean exists(Integer id) {
-        return wordDAO.exists(id);
+
+    public boolean exists(Word w) {
+        return wordDAO.exists(w);
     }
 }
